Day 24 - Get all working combos of size K.
Since the first bucket has to be as small as possible, calculate the
smallest possible bucket size and see if you can make the rest work.
If not, try the next smallest. This will be much more efficient.

First thing we need is the ability to find all valid combos of size K.